Major project files reorganization and renaming in some directories and files
