+++ lib/functionality_n_scripts/configuration_scripts/colort_list_invoke.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 import 'package:flutter/services.dart';
-import 'package:simple_chat/utils/logger_config.dart';
+import 'package:simple_chat/functionality_n_scripts/utils/logger_config.dart';
 
 //Globally accessible map
 Map<String, String> color_name_to_hex_map = {};
