+++ lib/functionality_n_scripts/session_related/refresh_tk_watch_dog.dart
@@ -1,10 +1,9 @@
 import 'dart:async';
 import 'package:flutter/cupertino.dart';
-import 'package:simple_chat/session_related/app_storage_class.dart';
-import 'package:simple_chat/configuration/config_invoke.dart';
-import 'package:simple_chat/utils/logger_config.dart';
-
-import 'package:simple_chat/standalone_methods_functions/user_entrypoint_methods.dart';
+import 'package:simple_chat/functionality_n_scripts/session_related/app_storage_class.dart';
+import 'package:simple_chat/functionality_n_scripts/configuration_scripts/config_invoke.dart';
+import 'package:simple_chat/functionality_n_scripts/utils/logger_config.dart';
+import 'package:simple_chat/functionality_n_scripts/standalone_methods/user_entrypoint_methods.dart';
 
 class TokenWatchdog {
   static final TokenWatchdog _instance = TokenWatchdog._internal();
