+++ lib/functionality_n_scripts/standalone_methods/general_methods.dart
@@ -11,11 +11,11 @@ import 'package:archive/archive.dart';
 import 'package:archive/archive_io.dart';
 
 //Import alert dialogs and others
-import "package:simple_chat/utils/widgets_and_ui_elements/alert_dialog_list.dart";
-import 'package:simple_chat/configuration/config_invoke.dart';
-import "package:simple_chat/message_related/message_class.dart";
-import 'package:simple_chat/configuration/colort_list_invoke.dart';
-import 'package:simple_chat/utils/logger_config.dart';
+import "package:simple_chat/widgets_and_ui_elements/alert_dialog_list.dart";
+import 'package:simple_chat/functionality_n_scripts/configuration_scripts/config_invoke.dart';
+import "package:simple_chat/functionality_n_scripts/message_related/message_class.dart";
+import 'package:simple_chat/functionality_n_scripts/configuration_scripts/colort_list_invoke.dart';
+import 'package:simple_chat/functionality_n_scripts/utils/logger_config.dart';
 
 //Audio instance
 final AudioPlayer _audio_instance = AudioPlayer();
