+++ lib/functionality_n_scripts/standalone_methods/user_entrypoint_methods.dart
@@ -5,13 +5,13 @@ import "package:flutter/material.dart";
 import "dart:async";
 import "dart:convert";
 import 'package:http/http.dart' as http;
-import "package:simple_chat/session_related/app_storage_class.dart";
-import "package:simple_chat/standalone_methods_functions/general_methods.dart";
+import "package:simple_chat/functionality_n_scripts/session_related/app_storage_class.dart";
+import "package:simple_chat/functionality_n_scripts/standalone_methods/general_methods.dart";
 
 //Import alert dialogs and others
-import "package:simple_chat/utils/widgets_and_ui_elements/alert_dialog_list.dart";
-import 'package:simple_chat/configuration/config_invoke.dart';
-import 'package:simple_chat/utils/logger_config.dart';
+import "package:simple_chat/widgets_and_ui_elements/alert_dialog_list.dart";
+import 'package:simple_chat/functionality_n_scripts/configuration_scripts/config_invoke.dart';
+import 'package:simple_chat/functionality_n_scripts/utils/logger_config.dart';
 
 //imports
 /////////////////////////////////////////////////////////////////////////////
