+++ lib/functionality_n_scripts/utils/easter_egg_player.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:audioplayers/audioplayers.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:simple_chat/configuration/config_invoke.dart';
+import 'package:simple_chat/functionality_n_scripts/configuration_scripts/config_invoke.dart';
 
 class EasterEggPlayerInline extends StatefulWidget {
   final String asset_path;
