+++ lib/screens_pages/log_in_page.dart
@@ -3,17 +3,17 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';   //Fonts
 
-import 'package:simple_chat/standalone_methods_functions/general_methods.dart';
-import 'package:simple_chat/configuration/config_invoke.dart';
+import 'package:simple_chat/functionality_n_scripts/standalone_methods/general_methods.dart';
+import 'package:simple_chat/functionality_n_scripts/configuration_scripts/config_invoke.dart';
 import 'package:simple_chat/screens_pages/landing_page.dart';
 import 'package:simple_chat/screens_pages/complete_profile_page.dart';
-import 'package:simple_chat/session_related/app_storage_class.dart';
-import 'package:simple_chat/utils/logger_config.dart';
+import 'package:simple_chat/functionality_n_scripts/session_related/app_storage_class.dart';
+import 'package:simple_chat/functionality_n_scripts/utils/logger_config.dart';
 import 'package:simple_chat/screens_pages/sign_up_page.dart';
 
-import 'package:simple_chat/session_related/refresh_tk_watch_dog.dart';
-import 'package:simple_chat/standalone_methods_functions/user_entrypoint_methods.dart';
-import 'package:simple_chat/utils/widgets_and_ui_elements/labeled_text_field.dart';
+import 'package:simple_chat/functionality_n_scripts/session_related/refresh_tk_watch_dog.dart';
+import 'package:simple_chat/functionality_n_scripts/standalone_methods/user_entrypoint_methods.dart';
+import 'package:simple_chat/widgets_and_ui_elements/labeled_text_field.dart';
 
 //imports
 /////////////////////////////////////////////////////////////////////////////
